rename `src/components/transitions/` to `src/components/transition/` (singular)
This is so that we can be consistent with the other components.
