+++ packages/@headlessui-react/src/hooks/use-transition.ts
@@ -1,5 +1,5 @@
 import { type MutableRefObject } from 'react'
-import { transition } from '../components/transitions/utils/transition'
+import { transition } from '../components/transition/utils/transition'
 import { disposables } from '../utils/disposables'
 import { useDisposables } from './use-disposables'
 import { useIsMounted } from './use-is-mounted'
